+++ CHANGELOG.md
@@ -20,6 +20,7 @@ All notable changes to this project will be documented in this file.
 - Support Xcode 7.3 and Swift 2.2
 - Configurable mask polygon (sides) [#112](https://github.com/JakeLin/IBAnimatable/issues/112)
 - Add `SystemPageCurlAnimator` to support `SystemPageCurlFromTop` and `SystemPageCurlFromBottom` transition animations [#126](https://github.com/JakeLin/IBAnimatable/issues/126)
+- `AnimatableLabel` now conforms `BorderDesignable` and `FillDesignable`
 
 #### Bugfixes